djangomission/requirements.txt
asgiref==3.2.10
black==22.8.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==3.1.14
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.5
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.1
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
virtualenv==20.16.5
